requirements.txt
opencv-python>=4.8.0
python-dotenv>=1.0.0
inference-sdk>=0.9.0
PySide6>=6.5.0
numpy>=1.24.0
requests>=2.31.0
